% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{getCorrUnad}
\alias{getCorrUnad}
\title{getCorrUnad}
\usage{
getCorrUnad(geneName, fp, fpName, responseVector, dat, resPlot)
}
\arguments{
\item{geneName}{Character string, specifying the name of the gene.}

\item{fp}{Vector containing 0's and 1's - the data about the fingerprint feature.}

\item{fpName}{Character string, used to make the title of the plots. If not specified, the plot title will be blank.}

\item{responseVector}{Vector containing the bio-activity data.}

\item{dat}{Contains the gene expression data matrix for all the genes - can be a matrix or an expression set.}

\item{resPlot}{Logical. If TRUE, creates the plot data for the residual plot}
}
\value{
A list containing the data to create the respective plots and the unadjusted association between the gene expression and bio-activity data.
}
\description{
The getCorrUnad function is a support function for the function plot1gene.
}
\details{
Works as a support function for plot1gene.
}
\examples{
\dontrun{
getCorrUnad(geneName="Gene21",fp=fp,fpName="Fingerprint",
responseVector=activity,dat=gene_eset,resPlot=TRUE)
}
}

