% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/IntegratedJM.R
\docType{package}
\name{IntegratedJM}
\alias{IntegratedJM}
\alias{IntegratedJM-package}
\title{Joint Modelling of Gene Expression and Bio-activity data, taking care of the effect due to Fingerprint Feature.}
\description{
The IntegratedJM package contains the functions to fit the Joint Model, to classify the genes based on different criteria and necessary plot functions.
}

