% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{multiplot}
\alias{multiplot}
\title{multiplot}
\usage{
multiplot(..., cols = 1)
}
\arguments{
\item{...}{ggplot2 objects, separated by comma.}

\item{cols}{Integer, specifying the number of plots in one row in the layout.}
}
\value{
Creates multiple ggplots in same window
}
\description{
The multiplot function plots multiple ggplots in the same window.
}
\details{
Plots multiple ggplots in the same window - multiplot(p1,p2,p3,p4, cols=2) is similar to the standard R notation par(mfrow=c(2,2)).
}
\examples{
\dontrun{
multiplot(p1,p2,p3,cols=3)
}
}

