% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{plot1gene}
\alias{plot1gene}
\title{plot1gene}
\usage{
plot1gene(geneName, fp, fpName = "", responseVector, dat, resPlot = TRUE,
  colP = "blue", colA = "white")
}
\arguments{
\item{geneName}{Character string, specifying the name of the gene.}

\item{fp}{Vector containing 0's and 1's - the data about the fingerprint feature.}

\item{fpName}{Character string, used to make the title of the plots. If not specified, the plot title will be blank.}

\item{responseVector}{Vector containing the bio-activity data.}

\item{dat}{Contains the gene expression data matrix for all the genes - can be a matrix or an expression set.}

\item{resPlot}{Logical. If TRUE, also plots the residual from the gls fit. Default is TRUE.}

\item{colP}{Character string, specifying the colour for the 1's in the fp parameter. Default is blue.}

\item{colA}{Character string, specifying the colour for the 0's in the fp parameter. Default is white.}
}
\value{
Creates a plot
}
\description{
The plot1gene function plots the data for a single gene.
}
\details{
Calls the getCorrUnad function and creates the plot(s) accordingly.
}
\examples{
\dontrun{
plot1gene(geneName="Gene21",fp=fp,fpName="Fingerprint",responseVector=activity,dat=gene_eset)
}
}

