% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Functions.R
\name{plotAsso}
\alias{plotAsso}
\title{plotAsso}
\usage{
plotAsso(jointModelResult, type)
}
\arguments{
\item{jointModelResult}{Data frame, containing the results from the fitJM function.}

\item{type}{Character string, specifying the type of association - Pearson or Spearman.}
}
\value{
Creates a plot
}
\description{
The plotAsso function is used to plot the unadjusted association vs the adjusted association for all the genes.
}
\details{
Plots the unadjusted association vs the adjusted association for all the genes.
}
\examples{
\dontrun{
plotAsso(jointModelResult=jmRes,type="Pearson")
}
}

